Move AgentPhoneNumberCollectionViewCell -> ContactPersonLinkCollectionViewCell
